+++ CHANGES.md
@@ -20,10 +20,8 @@
 ##### Fixes :wrench:
 
 - Fixed a bug that caused the `center` field of `AxisAlignedBox` to be incorrect.
-
-##### Fixes :wrench:
-
 - Fixed a bug that caused the main thread to sometimes load low-priority tiles before high-priority ones. This could result in much longer waits than necessary for a tileset's appropriate level-of-detail to be shown.
+- Fixed a bug that prevented WebP and KTX2 textures from working in the common case where only the extension specified the `source` property, not the glTF's main `Texture` definition.
 
 ### v0.22.1 - 2023-03-06